backend/requirements.txt
google-generativeai==0.3.1
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
slowapi==0.1.8
pytest==7.4.2
httpx==0.24.1